Added scripts to handle signature files
omfLibraryToSignatureList.py : processes a Watcom OMF library file and
emits a function name plus regex match pattern list that could be used
by the second script. The signature files need to be created only once
and could be shared even for each individual Watcom release.

mapSymbolsToDisassembly.py : sweeps a le_disasm disassembly file using
the signature list from the previous script and emits a list of symbol
to linear address pairs.

The generated list is not meant to be 100% accurate. The users need to
cross verify the signature matches whether they are accurate in which
case the found library functions could be added to the map file that
could be fed back into le_disasm to emit proper symbol names and more
accurate assembly code in general.